+++ src/actionTopicWrite.cpp
@@ -53,12 +53,12 @@ execution ActionTopicWrite::execute(const TestRepetitions& testrepetition)
 		{
 			auto publisher = create_publisher<geometry_msgs::msg::Twist>(topic_, 10);
 			auto message = geometry_msgs::msg::Twist();
-			int x = j.at(rossyntax::dataTypeGeometryTwist).value("x",0);
-			int y = j.at(rossyntax::dataTypeGeometryTwist).value("y",0);
-			int z = j.at(rossyntax::dataTypeGeometryTwist).value("z",0);
-			int xa = j.at(rossyntax::dataTypeGeometryTwist).value("xa",0);
-			int ya = j.at(rossyntax::dataTypeGeometryTwist).value("ya",0);
-			int za = j.at(rossyntax::dataTypeGeometryTwist).value("za",0);
+			float x = j.at(rossyntax::dataTypeGeometryTwist).value("x",0);
+			float y = j.at(rossyntax::dataTypeGeometryTwist).value("y",0);
+			float z = j.at(rossyntax::dataTypeGeometryTwist).value("z",0);
+			float xa = j.at(rossyntax::dataTypeGeometryTwist).value("xa",0);
+			float ya = j.at(rossyntax::dataTypeGeometryTwist).value("ya",0);
+			float za = j.at(rossyntax::dataTypeGeometryTwist).value("za",0);
 
 			message.angular.x = xa;
 			message.angular.y = ya;
